configure.ac - filter -L/usr/lib from HDF5_LDFLAGS
HDF5 is configured via AX_LIB_HDF5. This finds build flags from the
output of 'h5cc -show'. On DiRAC COSMOS, this includes -L/usr/lib64.
This is not necessary for the SWIFT build, and also means that the
system copy of METIS (which is not recent enough) is found before a
user-built copy, even with a correctly-configured environment.

As a 'fix', filter anything which looks like -L/usr/lib(32|64)? from
HDF5_LDFLAGS after AX_LIB_HDF5 returns. I don't think it's impossible
that this will itself break under unusual environments, but it seems
fairly safe: COSMA, Complexity, COSMOS and HPCS are OK with the change.